requirements.txt
aws-cdk-lib==2.66.0
constructs==10.1.257
